% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSTARCounts.R
\name{loadSTARCounts}
\alias{loadSTARCounts}
\title{Load STAR aligner gene count output}
\usage{
loadSTARCounts(fnames, quant = c("unstranded", "FRaligned", "SRaligned"))
}
\arguments{
\item{fnames}{The file names for each of the .tab files from STAR.}

\item{quant}{Which column from the STAR output to use. See details section
for more information.}
}
\value{
A matrix of counts, with genes from the STAR output on the rownames
and file names on the columns.
}
\description{
This function assumes that STAR aligner has been run with the
\code{--quantMode GeneCounts} option. It simply reads in the data
files, one-per-sample, into a matrix with rownames corresponding
to the gene names in the STAR output and column names that are
the file names passed to the function as the first parameter.
}
\details{
The choice of \code{quant} is based on how STAR interprets the
reads relative to strandedness of the protocol used to produce
the data.  In particular, \code{unstranded} is used for an
unstranded protocol, \code{FRaligned} is for when the first read
is in the same direction as the strand of the mRNA, and
\code{SRaligned} is for when the second read is in the same
direction as the strand of the mRNA.
}
\examples{
fnames = dir(system.file(package='SeansStuff','extdata'),full.names=TRUE)
mat = loadSTARCounts(fnames)
dim(mat)
head(mat)
# clean up column names
colnames(mat) = basename(colnames(mat))
head(mat)
}

