% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myHm.R
\name{myHm}
\alias{myHm}
\title{My heatmap}
\usage{
myHm(eset, n, ...)
}
\arguments{
\item{eset}{An ExpressionSet}

\item{n}{The top most variable genes}

\item{...}{passed to heatmap}
}
\description{
This is a cool heatmap function
}

